+++ src/vs/workbench/contrib/extensions/browser/extensionRecommendationsService.ts
@@ -4,11 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { localize } from 'vs/nls';
-import { forEach, groupBy } from 'vs/base/common/collections';
+import { forEach, IStringDictionary } from 'vs/base/common/collections';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { match } from 'vs/base/common/glob';
 import * as json from 'vs/base/common/json';
-import { IExtensionManagementService, IExtensionGalleryService, EXTENSION_IDENTIFIER_PATTERN, InstallOperation, ILocalExtension } from 'vs/platform/extensionManagement/common/extensionManagement';
+import { IExtensionManagementService, IExtensionGalleryService, EXTENSION_IDENTIFIER_PATTERN, InstallOperation, ILocalExtension, IExecutableBasedExtensionTip, IExtensionTipsService } from 'vs/platform/extensionManagement/common/extensionManagement';
 import { IExtensionRecommendationsService, ExtensionRecommendationReason, IExtensionsConfigContent, RecommendationChangeNotification, IExtensionRecommendation, ExtensionRecommendationSource, IWorkbenchExtensionEnablementService, EnablementState } from 'vs/workbench/services/extensionManagement/common/extensionManagement';
 import { IModelService } from 'vs/editor/common/services/modelService';
 import { ITextModel } from 'vs/editor/common/model';
@@ -16,12 +16,12 @@ import { IStorageService, StorageScope } from 'vs/platform/storage/common/storag
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { ShowRecommendedExtensionsAction, InstallWorkspaceRecommendedExtensionsAction, InstallRecommendedExtensionAction } from 'vs/workbench/contrib/extensions/browser/extensionsActions';
 import Severity from 'vs/base/common/severity';
-import { IWorkspaceContextService, IWorkspaceFolder, IWorkspace, IWorkspaceFoldersChangeEvent } from 'vs/platform/workspace/common/workspace';
+import { IWorkspaceContextService, IWorkspaceFolder, IWorkspace, IWorkspaceFoldersChangeEvent, WorkbenchState } from 'vs/platform/workspace/common/workspace';
 import { IFileService } from 'vs/platform/files/common/files';
-import { IExtensionsConfiguration, ConfigurationKey, IExtensionsViewPaneContainer, IExtensionsWorkbenchService, EXTENSIONS_CONFIG } from 'vs/workbench/contrib/extensions/common/extensions';
+import { IExtensionsConfiguration, ConfigurationKey, IExtensionsViewPaneContainer, IExtensionsWorkbenchService, EXTENSIONS_CONFIG, ShowRecommendationsOnlyOnDemandKey } from 'vs/workbench/contrib/extensions/common/extensions';
 import { IConfigurationService, ConfigurationTarget } from 'vs/platform/configuration/common/configuration';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
-import { flatten, distinct, coalesce } from 'vs/base/common/arrays';
+import { flatten, distinct, coalesce, isNonEmptyArray, shuffle } from 'vs/base/common/arrays';
 import { guessMimeTypes, MIME_UNKNOWN } from 'vs/base/common/mime';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 import { IViewletService } from 'vs/workbench/services/viewlet/browser/viewlet';
@@ -38,8 +38,13 @@ import { IProductService } from 'vs/platform/product/common/productService';
 import { setImmediate } from 'vs/base/common/platform';
 import { Schemas } from 'vs/base/common/network';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
+import { timeout } from 'vs/base/common/async';
+import { LifecyclePhase, ILifecycleService } from 'vs/platform/lifecycle/common/lifecycle';
+import { IWorkspaceTagsService } from 'vs/workbench/contrib/tags/common/workspaceTags';
+import { isNumber } from 'vs/base/common/types';
+import { basename } from 'vs/base/common/path';
 
-export type ExtensionRecommendationsNotificationClassification = {
+type ExtensionRecommendationsNotificationClassification = {
 	userReaction: { classification: 'SystemMetaData', purpose: 'FeatureInsight' };
 	extensionId: { classification: 'PublicNonPersonalData', purpose: 'FeatureInsight' };
 };
@@ -53,7 +58,20 @@ type ExtensionWorkspaceRecommendationsNotificationClassification = {
 	userReaction: { classification: 'SystemMetaData', purpose: 'FeatureInsight' };
 };
 
-export const milliSecondsInADay = 1000 * 60 * 60 * 24;
+type DynamicWorkspaceRecommendationsClassification = {
+	count: { classification: 'SystemMetaData', purpose: 'FeatureInsight', isMeasurement: true };
+	cache: { classification: 'SystemMetaData', purpose: 'FeatureInsight', isMeasurement: true };
+};
+
+type ExeExtensionRecommendationsClassification = {
+	extensionId: { classification: 'PublicNonPersonalData', purpose: 'FeatureInsight' };
+	exeName: { classification: 'PublicNonPersonalData', purpose: 'FeatureInsight' };
+};
+
+type IStoredDynamicWorkspaceRecommendations = { recommendations: string[], timestamp: number };
+
+const dynamicWorkspaceRecommendationsStorageKey = 'extensionsAssistant/dynamicWorkspaceRecommendations';
+const milliSecondsInADay = 1000 * 60 * 60 * 24;
 const choiceNever = localize('neverShowAgain', "Don't Show Again");
 const searchMarketplace = localize('searchMarketplace', "Search Marketplace");
 const processedFileExtensions: string[] = [];
@@ -79,6 +97,8 @@ export class ExtensionRecommendationsService extends Disposable implements IExte
 	private availableRecommendations: { [pattern: string]: string[] } = Object.create(null);
 	private allWorkspaceRecommendedExtensions: IExtensionRecommendation[] = [];
 	private experimentalRecommendations: { [id: string]: string } = Object.create(null);
+	private exeBasedRecommendations: { [id: string]: IExecutableBasedExtensionTip; } = Object.create(null);
+	private dynamicWorkspaceRecommendations: string[] = [];
 
 	// Ignored Recommendations
 	private allIgnoredRecommendations: string[] = [];
@@ -86,42 +106,57 @@ export class ExtensionRecommendationsService extends Disposable implements IExte
 	private workspaceIgnoredRecommendations: string[] = [];
 
 	public loadWorkspaceConfigPromise: Promise<void>;
+	private sessionSeed: number;
 
 	private readonly _onRecommendationChange = this._register(new Emitter<RecommendationChangeNotification>());
 	onRecommendationChange: Event<RecommendationChangeNotification> = this._onRecommendationChange.event;
 
 	constructor(
-		@IExtensionGalleryService protected readonly galleryService: IExtensionGalleryService,
+		@IExtensionGalleryService private readonly galleryService: IExtensionGalleryService,
 		@IModelService private readonly modelService: IModelService,
-		@IStorageService protected readonly storageService: IStorageService,
+		@IStorageService private readonly storageService: IStorageService,
 		@IExtensionManagementService private readonly extensionsService: IExtensionManagementService,
 		@IWorkbenchExtensionEnablementService private readonly extensionEnablementService: IWorkbenchExtensionEnablementService,
-		@IInstantiationService protected readonly instantiationService: IInstantiationService,
-		@IFileService protected readonly fileService: IFileService,
-		@IWorkspaceContextService protected readonly contextService: IWorkspaceContextService,
-		@IConfigurationService protected readonly configurationService: IConfigurationService,
-		@ITelemetryService protected readonly telemetryService: ITelemetryService,
-		@IEnvironmentService protected readonly environmentService: IEnvironmentService,
+		@IInstantiationService private readonly instantiationService: IInstantiationService,
+		@IFileService private readonly fileService: IFileService,
+		@IWorkspaceContextService private readonly contextService: IWorkspaceContextService,
+		@IConfigurationService private readonly configurationService: IConfigurationService,
+		@ITelemetryService private readonly telemetryService: ITelemetryService,
+		@IEnvironmentService private readonly environmentService: IEnvironmentService,
 		@IExtensionService private readonly extensionService: IExtensionService,
 		@IViewletService private readonly viewletService: IViewletService,
-		@INotificationService protected readonly notificationService: INotificationService,
-		@IExtensionManagementService protected readonly extensionManagementService: IExtensionManagementService,
+		@INotificationService private readonly notificationService: INotificationService,
+		@IExtensionManagementService private readonly extensionManagementService: IExtensionManagementService,
 		@IExtensionsWorkbenchService private readonly extensionWorkbenchService: IExtensionsWorkbenchService,
 		@IExperimentService private readonly experimentService: IExperimentService,
-		@IProductService protected readonly productService: IProductService
+		@IProductService private readonly productService: IProductService,
+		@IWorkspaceTagsService private readonly workspaceTagsService: IWorkspaceTagsService,
+		@IExtensionTipsService private readonly extensionTipsService: IExtensionTipsService,
+		@ILifecycleService private readonly lifecycleService: ILifecycleService,
 	) {
 		super();
 
 		if (!this.isEnabled()) {
+			this.sessionSeed = 0;
 			this.loadWorkspaceConfigPromise = Promise.resolve();
 			return;
 		}
+
+		this.sessionSeed = +new Date();
 
 		const globallyIgnored = <string[]>JSON.parse(this.storageService.get('extensionsAssistant/ignored_recommendations', StorageScope.GLOBAL, '[]'));
 		this.globallyIgnoredRecommendations = globallyIgnored.map(id => id.toLowerCase());
 
 		this.fetchFileBasedRecommendations();
 		this.fetchExperimentalRecommendations();
+
+		/* 3s has come out to be the good number to fetch and prompt important exe based recommendations */
+		/* Fetch important exe based recommendations always for reporting telemetry */
+		timeout(3000).then(() => this.fetchAndPromptImportantExeBasedRecommendations());
+
+		if (!this.configurationService.getValue<boolean>(ShowRecommendationsOnlyOnDemandKey)) {
+			this.lifecycleService.when(LifecyclePhase.Eventually).then(() => this.fetchProactiveRecommendations());
+		}
 
 		this.loadWorkspaceConfigPromise = this.getWorkspaceRecommendations().then(() => {
 			this.promptWorkspaceRecommendations();
@@ -150,12 +185,35 @@ export class ExtensionRecommendationsService extends Disposable implements IExte
 		}));
 	}
 
-	protected isEnabled(): boolean {
+	private isEnabled(): boolean {
 		return this.galleryService.isEnabled() && !this.environmentService.extensionDevelopmentLocationURI;
 	}
 
 	getAllRecommendationsWithReason(): { [id: string]: { reasonId: ExtensionRecommendationReason, reasonText: string }; } {
+		/* Trigger fetching proactive recommendations */
+		this.fetchProactiveRecommendations();
+
 		const output: { [id: string]: { reasonId: ExtensionRecommendationReason, reasonText: string }; } = Object.create(null);
+
+		if (this.contextService.getWorkspace().folders && this.contextService.getWorkspace().folders.length === 1) {
+			const currentRepo = this.contextService.getWorkspace().folders[0].name;
+
+			this.dynamicWorkspaceRecommendations.forEach(id => output[id.toLowerCase()] = {
+				reasonId: ExtensionRecommendationReason.DynamicWorkspace,
+				reasonText: localize('dynamicWorkspaceRecommendation', "This extension may interest you because it's popular among users of the {0} repository.", currentRepo)
+			});
+		}
+
+		forEach(this.exeBasedRecommendations, entry => output[entry.key.toLowerCase()] = {
+			reasonId: ExtensionRecommendationReason.Executable,
+			reasonText: localize('exeBasedRecommendation', "This extension is recommended because you have {0} installed.", entry.value.friendlyName)
+		});
+
+		for (const id of Object.keys(output)) {
+			if (!this.isExtensionAllowedToBeRecommended(id)) {
+				delete output[id];
+			}
+		}
 
 		forEach(this.experimentalRecommendations, entry => output[entry.key.toLowerCase()] = {
 			reasonId: ExtensionRecommendationReason.Experimental,
@@ -676,7 +734,7 @@ export class ExtensionRecommendationsService extends Disposable implements IExte
 		);
 	}
 
-	protected promptExtensionInstallNotification(extensionId: string, message: string): void {
+	private promptExtensionInstallNotification(extensionId: string, message: string): void {
 		this.notificationService.prompt(Severity.Info, message,
 			[{
 				label: localize('install', 'Install'),
@@ -721,7 +779,7 @@ export class ExtensionRecommendationsService extends Disposable implements IExte
 		);
 	}
 
-	protected filterIgnoredOrNotAllowed(recommendationsToSuggest: string[]): string[] {
+	private filterIgnoredOrNotAllowed(recommendationsToSuggest: string[]): string[] {
 		const importantRecommendationsIgnoreList = <string[]>JSON.parse(this.storageService.get('extensionsAssistant/importantRecommendationsIgnore', StorageScope.GLOBAL, '[]'));
 		return recommendationsToSuggest.filter(id => {
 			if (importantRecommendationsIgnoreList.indexOf(id) !== -1) {
@@ -739,13 +797,20 @@ export class ExtensionRecommendationsService extends Disposable implements IExte
 		return recommendationsToSuggest.filter(id => !installedExtensionsIds.has(id.toLowerCase()));
 	}
 
-	protected groupByInstalled(recommendationsToSuggest: string[], local: ILocalExtension[]): { installed: string[], uninstalled: string[] } {
+	private groupByInstalled(recommendationsToSuggest: string[], local: ILocalExtension[]): { installed: string[], uninstalled: string[] } {
+		const installed: string[] = [], uninstalled: string[] = [];
 		const installedExtensionsIds = local.reduce((result, i) => { result.add(i.identifier.id.toLowerCase()); return result; }, new Set<string>());
-		const result = groupBy(recommendationsToSuggest, id => installedExtensionsIds.has(id.toLowerCase()) ? 'installed' : 'uninstalled');
-		return result as { installed: string[], uninstalled: string[] };
+		recommendationsToSuggest.forEach(id => {
+			if (installedExtensionsIds.has(id.toLowerCase())) {
+				installed.push(id);
+			} else {
+				uninstalled.push(id);
+			}
+		});
+		return { installed, uninstalled };
 	}
 
-	protected addToImportantRecommendationsIgnore(id: string) {
+	private addToImportantRecommendationsIgnore(id: string) {
 		const importantRecommendationsIgnoreList = <string[]>JSON.parse(this.storageService.get('extensionsAssistant/importantRecommendationsIgnore', StorageScope.GLOBAL, '[]'));
 		importantRecommendationsIgnoreList.push(id);
 		this.storageService.store(
@@ -755,7 +820,7 @@ export class ExtensionRecommendationsService extends Disposable implements IExte
 		);
 	}
 
-	protected setIgnoreRecommendationsConfig(configVal: boolean) {
+	private setIgnoreRecommendationsConfig(configVal: boolean) {
 		this.configurationService.updateValue('extensions.ignoreRecommendations', configVal, ConfigurationTarget.USER);
 		if (configVal) {
 			const ignoreWorkspaceRecommendationsStorageKey = 'extensionsAssistant/workspaceRecommendationsIgnore';
@@ -768,9 +833,26 @@ export class ExtensionRecommendationsService extends Disposable implements IExte
 	//#region otherRecommendations
 
 	async getOtherRecommendations(): Promise<IExtensionRecommendation[]> {
-		return Object.keys(this.experimentalRecommendations)
+		const otherRecommendations = Object.keys(this.experimentalRecommendations)
 			.filter(extensionId => this.isExtensionAllowedToBeRecommended(extensionId))
 			.map(extensionId => (<IExtensionRecommendation>{ extensionId, sources: [] }));
+		await this.fetchProactiveRecommendations();
+		const others = distinct([
+			...Object.keys(this.exeBasedRecommendations),
+			...this.dynamicWorkspaceRecommendations,
+			...otherRecommendations.map(e => e.extensionId),
+		]).filter(extensionId => this.isExtensionAllowedToBeRecommended(extensionId));
+		shuffle(others, this.sessionSeed);
+		return others.map(extensionId => {
+			const sources: ExtensionRecommendationSource[] = [];
+			if (this.exeBasedRecommendations[extensionId]) {
+				sources.push('executable');
+			}
+			if (this.dynamicWorkspaceRecommendations.indexOf(extensionId) !== -1) {
+				sources.push('dynamic');
+			}
+			return (<IExtensionRecommendation>{ extensionId, sources });
+		});
 	}
 
 	/**
@@ -791,11 +873,126 @@ export class ExtensionRecommendationsService extends Disposable implements IExte
 
 	//#endregion
 
-	protected isExtensionAllowedToBeRecommended(id: string): boolean {
+	//#region proactive recommendations - exe based and dynamic workspace recommendations
+
+	private proactiveRecommendationsPromise: Promise<void> | undefined = undefined;
+	private async fetchProactiveRecommendations(): Promise<void> {
+		if (!this.proactiveRecommendationsPromise) {
+			this.proactiveRecommendationsPromise = Promise.all([this.fetchDynamicWorkspaceRecommendations(), this.fetchOtherExeBasedRecommendations()])
+				.then(() => {
+					this._register(this.contextService.onDidChangeWorkbenchState(() => this.dynamicWorkspaceRecommendations = []));
+				});
+		}
+		return this.proactiveRecommendationsPromise;
+	}
+
+	private async fetchAndPromptImportantExeBasedRecommendations(): Promise<void> {
+		const importantExeBasedRecommendations: IStringDictionary<IExecutableBasedExtensionTip> = {};
+		const importantExectuableBasedTips = await this.extensionTipsService.getImportantExecutableBasedTips();
+		importantExectuableBasedTips.forEach(tip => {
+			this.exeBasedRecommendations[tip.extensionId.toLowerCase()] = tip;
+			importantExeBasedRecommendations[tip.extensionId.toLowerCase()] = tip;
+		});
+
+		const local = await this.extensionManagementService.getInstalled(ExtensionType.User);
+		const { installed, uninstalled } = this.groupByInstalled(Object.keys(importantExeBasedRecommendations), local);
+
+		/* Log installed and uninstalled exe based recommendations */
+		for (const extensionId of installed) {
+			const tip = importantExeBasedRecommendations[extensionId];
+			this.telemetryService.publicLog2<{ exeName: string, extensionId: string }, ExeExtensionRecommendationsClassification>('exeExtensionRecommendations:alreadyInstalled', { extensionId, exeName: basename(tip.windowsPath!) });
+		}
+		for (const extensionId of uninstalled) {
+			const tip = importantExeBasedRecommendations[extensionId];
+			this.telemetryService.publicLog2<{ exeName: string, extensionId: string }, ExeExtensionRecommendationsClassification>('exeExtensionRecommendations:notInstalled', { extensionId, exeName: basename(tip.windowsPath!) });
+		}
+
+		this.promptImportantExeBasedRecommendations(uninstalled, importantExeBasedRecommendations);
+	}
+
+	private promptImportantExeBasedRecommendations(recommendations: string[], importantExeBasedRecommendations: IStringDictionary<IExecutableBasedExtensionTip>): void {
+		if (this.hasToIgnoreRecommendationNotifications()) {
+			return;
+		}
+		recommendations = this.filterIgnoredOrNotAllowed(recommendations);
+		if (recommendations.length === 0) {
+			return;
+		}
+
+		const extensionId = recommendations[0];
+		const tip = importantExeBasedRecommendations[extensionId];
+		const message = localize('exeRecommended', "The '{0}' extension is recommended as you have {1} installed on your system.", tip.friendlyName!, tip.exeFriendlyName || basename(tip.windowsPath!));
+		this.promptExtensionInstallNotification(extensionId, message);
+	}
+
+	private async fetchOtherExeBasedRecommendations(): Promise<void> {
+		const otherExectuableBasedTips = await this.extensionTipsService.getOtherExecutableBasedTips();
+		otherExectuableBasedTips.forEach(tip => this.exeBasedRecommendations[tip.extensionId.toLowerCase()] = tip);
+	}
+
+	/**
+	 * Fetch extensions used by others on the same workspace as recommendations
+	 */
+	private async fetchDynamicWorkspaceRecommendations(): Promise<void> {
+		if (this.dynamicWorkspaceRecommendations.length
+			|| this.contextService.getWorkbenchState() !== WorkbenchState.FOLDER
+			|| !this.fileService.canHandleResource(this.contextService.getWorkspace().folders[0].uri)
+		) {
+			return;
+		}
+
+		const cachedDynamicWorkspaceRecommendations = this.getCachedDynamicWorkspaceRecommendations();
+		if (cachedDynamicWorkspaceRecommendations) {
+			this.dynamicWorkspaceRecommendations = cachedDynamicWorkspaceRecommendations;
+			this.telemetryService.publicLog2<{ count: number, cache: number }, DynamicWorkspaceRecommendationsClassification>('dynamicWorkspaceRecommendations', { count: this.dynamicWorkspaceRecommendations.length, cache: 1 });
+			return;
+		}
+
+		const workspaceUri = this.contextService.getWorkspace().folders[0].uri;
+		const [hashedRemotes1, hashedRemotes2] = await Promise.all([this.workspaceTagsService.getHashedRemotesFromUri(workspaceUri, false), this.workspaceTagsService.getHashedRemotesFromUri(workspaceUri, true)]);
+		const hashedRemotes = (hashedRemotes1 || []).concat(hashedRemotes2 || []);
+		if (!hashedRemotes.length) {
+			return;
+		}
+
+		const workspacesTips = await this.extensionTipsService.getAllWorkspacesTips();
+		if (!workspacesTips.length) {
+			return;
+		}
+
+		for (const hashedRemote of hashedRemotes) {
+			const workspaceTip = workspacesTips.filter(workspaceTip => isNonEmptyArray(workspaceTip.remoteSet) && workspaceTip.remoteSet.indexOf(hashedRemote) > -1)[0];
+			if (workspaceTip) {
+				this.dynamicWorkspaceRecommendations = workspaceTip.recommendations.filter(id => this.isExtensionAllowedToBeRecommended(id)) || [];
+				this.storageService.store(dynamicWorkspaceRecommendationsStorageKey, JSON.stringify(<IStoredDynamicWorkspaceRecommendations>{ recommendations: this.dynamicWorkspaceRecommendations, timestamp: Date.now() }), StorageScope.WORKSPACE);
+				this.telemetryService.publicLog2<{ count: number, cache: number }, DynamicWorkspaceRecommendationsClassification>('dynamicWorkspaceRecommendations', { count: this.dynamicWorkspaceRecommendations.length, cache: 0 });
+				return;
+			}
+		}
+	}
+
+	private getCachedDynamicWorkspaceRecommendations(): string[] | undefined {
+		try {
+			const storedDynamicWorkspaceRecommendations: IStoredDynamicWorkspaceRecommendations = JSON.parse(this.storageService.get(dynamicWorkspaceRecommendationsStorageKey, StorageScope.WORKSPACE, '{}'));
+			if (isNonEmptyArray(storedDynamicWorkspaceRecommendations.recommendations)
+				&& isNumber(storedDynamicWorkspaceRecommendations.timestamp)
+				&& storedDynamicWorkspaceRecommendations.timestamp > 0
+				&& (Date.now() - storedDynamicWorkspaceRecommendations.timestamp) / milliSecondsInADay < 14) {
+				return storedDynamicWorkspaceRecommendations.recommendations;
+			}
+		} catch (e) {
+			this.storageService.remove(dynamicWorkspaceRecommendationsStorageKey, StorageScope.WORKSPACE);
+		}
+		return undefined;
+	}
+
+	//#endregion
+
+	private isExtensionAllowedToBeRecommended(id: string): boolean {
 		return this.allIgnoredRecommendations.indexOf(id.toLowerCase()) === -1;
 	}
 
-	protected hasToIgnoreRecommendationNotifications(): boolean {
+	private hasToIgnoreRecommendationNotifications(): boolean {
 		const config = this.configurationService.getValue<IExtensionsConfiguration>(ConfigurationKey);
 		return config.ignoreRecommendations || config.showRecommendationsOnlyOnDemand;
 	}
