+++ src/vs/workbench/workbench.web.main.ts
@@ -113,9 +113,6 @@ import 'vs/workbench/contrib/debug/browser/extensionHostDebugService';
 import 'vs/workbench/contrib/webview/browser/webviewService';
 import 'vs/workbench/contrib/webview/browser/webviewWorkbenchService';
 
-// Extensions Management
-import 'vs/workbench/contrib/extensions/browser/extensions.web.contribution';
-
 // Terminal
 import 'vs/workbench/contrib/terminal/browser/terminalInstanceService';
